numbers: Use uintmax_t for numlen

+++ include/libft/numbers.h
@@ -1,8 +1,10 @@
 #pragma once
 
+#include <stdint.h>
+
 char		*ft_itoa(int n);
 int			ft_snitoa(char *dest, int n, int size);
 
-int			ft_numlen(unsigned long long n, unsigned base);
+int			ft_numlen(uintmax_t n, unsigned base);
 long double	ft_pow(int nb, int pow);
 int			ft_atoi(const char *str);
